Convert `learner.py` into `learner` package
Split into `learner/__init__.py` and `learner/utils.py`.
